backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.8.2
google-generativeai>=0.7.2
httpx==0.27.2
python-dotenv==1.0.1
anyio==4.4.0
